Add ability to continue a transaction without a re-start

+++ src/imx_rt1060/imx_rt1060_i2c_driver.h
@@ -120,9 +120,9 @@ public:
 
     size_t get_bytes_transferred() override;
 
-    void write_async(uint8_t address, const uint8_t* buffer, size_t num_bytes, bool send_stop) override;
+    void write_async(int address, const uint8_t* buffer, size_t num_bytes, bool send_stop) override;
 
-    void read_async(uint8_t address, uint8_t* buffer, size_t num_bytes, bool send_stop) override;
+    void read_async(int address, uint8_t* buffer, size_t num_bytes, bool send_stop) override;
 
     // DO NOT call this method directly.
     void _interrupt_service_routine();
@@ -151,7 +151,7 @@ private:
     void (* isr)();
     void set_clock(uint32_t frequency);
     void abort_transaction_async();
-    bool start(uint8_t address, uint32_t direction);
+    bool start(int address, uint32_t direction);
     uint8_t tx_fifo_count();
     uint8_t rx_fifo_count();
     void clear_all_msr_flags();
